+++ packages/api/models/schemas.js
@@ -370,7 +370,6 @@ module.exports.granule = {
     'granuleId',
     'collectionId',
     'status',
-    'execution',
     'createdAt',
     'updatedAt',
   ],